Create Card component and add it to GameBoard
The Card component changes the shape displayed based on given props (by GameBoard).
Style was added to GameBoard and Card in order to make the display work.

Things to consider: With the way style was implemented, the ratios of the shape can change depending on the ratio of GameBoard, this can be problematic for shapes such as square and circle, hence, circle was changed to ellipse (no issues seen with square yet).